+++ releases.md
@@ -2,7 +2,14 @@
 title: Version history
 ---
 
-In planning stages only: no published versions.
 
 
-**0.1.0**, *?*: initial Marimo notebook.
+
+**Sept. 23, 2026**: 
+
+- adds display and filtering by editorial status
+- adds optional display of interpretive comments from LM
+- organizes display in three accordions with LM comments, formatted text, and diagrams
+
+
+**Sept. 22, 2026**: preliminary experimental notebook published at https://shot.holycross.edu/readers/proarchia.